// src/api/generated/models/authorShort.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API каталога книг
 * REST API для управления каталогом книг и авторов.
 * Поддерживает роли: гость (просмотр) и пользователь (CRUD).
 * Аутентификация через Bearer-токен (JWT).
 *
 * OpenAPI spec version: 1.0.0
 */

export interface AuthorShort {
  id?: number;
  full_name?: string;
}